Cambios menores

+++ dominio/src/main/java/promotionSystem/hechizo/IraWookeana.java
@@ -7,7 +7,7 @@ public class IraWookeana extends Hechizo{
 	@Override
 	public void afectar(Personaje personaje, int efecto) {
 		
-		int aumento=(int)(efecto/3);
+		int aumento= efecto/3;
 		
 		personaje.serAumentadoLaFuerza(aumento);
